// src/app/modals/model/current-user.ts
export class CurrentUser {
     userProfileImg: string;
     userDescription: string;
     userName: string;
     userEmail: string;
     userDateCreated: Date;
     userPhoneNo: string;
     userBaseLocation: string;
     userType: string;
     userService: string;
     userPosts: [];
     userReviews: [];
     userOrders: [];

     // constructor(
     //      userProfileImg: string,
     //      userDescription: string,
     //      userName: string,
     //      userEmail: string,
     //      userDateCreated: Date,
     //      userPhoneNo: string,
     //      userBaseLocation: string,
     //      userType: string,
     //      userService: string,
     //      userPosts: [],
     //      userReviews: [],
     //      userOrders: [],
     // ) {
     //      this.userProfileImg = userProfileImg;
     //      this.userDescription = userDescription;
     //      this.userName = userName;
     //      this.userEmail = userEmail;
     //      this.userDateCreated = userDateCreated;
     //      this.userPhoneNo = userPhoneNo;
     //      this.userBaseLocation = userBaseLocation;
     //      this.userType = userType;
     //      this.userService = userService;
     //      this.userPosts = userPosts;
     //      this.userReviews = userReviews;
     //      this.userOrders = userOrders;
     // }
}